Rename parser2 -> parser, nom5 -> nom, apply clippy.
