+++ Cargo.toml
@@ -6,9 +6,9 @@ edition = "2018"
 
 [dependencies]
 itertools = "0.8.0"
-rand = "0.7"
-rand_xorshift = "0.2"
-nom5 = { package = "nom", version = "5.0" }
+nom = "5.0.0"
+rand = "0.7.0"
+rand_xorshift = "0.2.0"
 
 [dev-dependencies]
 diff = "0.1.11"
