+++ src/lib.rs
@@ -2,7 +2,7 @@ mod ast;
 mod error;
 mod format;
 mod interpreter;
-mod parser2;
+mod parser;
 
 pub use crate::error::Error;
 use crate::interpreter::Interpreter;
@@ -15,7 +15,7 @@ pub fn execute<R: BufRead, W: Write, V: Write>(
     stdout: &mut W,
     stderr: &mut V,
 ) -> Result<(), Error> {
-    parser2::program(input).and_then(|ast| {
+    parser::program(input).and_then(|ast| {
         let interpreter = Interpreter::new(&ast, input);
         interpreter.evaluate(stdin, stdout, stderr)
     })
